feat: 实现SHEIN限时折扣活动自动报名和任务依赖管理系统
主要功能:
1. 限时折扣活动自动报名
   - 支持按折扣率和利润率两种定价模式
   - 自动过滤库存不足、已参加活动、折扣不足的商品
   - 支持活动库存限量配置
   - 最多500个商品/活动的限制
   - 完整的价格计算和风险检查

2. 任务依赖管理系统
   - 实现任务间的依赖关系管理
   - 支持任务执行顺序控制(产品同步库存监控活动报名)
   - 自动等待依赖任务完成
   - 任务执行状态跟踪和监控

3. 代码优化
   - 拆分activity_registration.go为多个职责单一的文件
   - 添加价格计算器模块
   - 完善错误处理和日志记录
   - 添加详细的文档说明

4. 文档
   - ACTIVITY_REGISTRATION.md: 活动报名功能说明
   - TASK_DEPENDENCY.md: 任务依赖管理说明
   - TIME_LIMITED_DISCOUNT_OPTIONS.md: 限时折扣配置选项

+++ internal/platforms/shein/api/marketing/price_types.go
@@ -44,7 +44,7 @@ type SkuCalculationInfo struct {
 	SkuCode      string    `json:"sku_code"`      // SKU编码
 	PriceInfo    PriceInfo `json:"price_info"`    // 价格信息
 	RiskTag      int       `json:"risk_tag"`      // 风险标签（0:无风险）
-	WarningValue int       `json:"warning_value"` // 警告值
+	WarningValue float64   `json:"warning_value"` // 警告值
 }
 
 // PriceInfo 价格详细信息
